+++ NEWS.md
@@ -1,3 +1,9 @@
+# RSAddins 1.1.1
+
+* Duplicate/delete section
+* Better insertDput for vectors
+
+
 # RSAddins 1.1.0
 
 * Many new addins and features :)